+++ instrumentation/apache-httpclient/apache-httpclient-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v2_0/ApacheHttpClientHttpAttributesExtractor.java
@@ -75,9 +75,9 @@ final class ApacheHttpClientHttpAttributesExtractor
 
   @Override
   @Nullable
-  protected Long statusCode(HttpMethod httpMethod, Void unused) {
+  protected Integer statusCode(HttpMethod httpMethod, Void unused) {
     StatusLine statusLine = httpMethod.getStatusLine();
-    return statusLine == null ? null : (long) statusLine.getStatusCode();
+    return statusLine == null ? null : statusLine.getStatusCode();
   }
 
   @Override
